First working version of the scrubber
This took me nearly 4 hours to do. I went through a couple of design choices but went with something that will always scan the same url in Kaggle, but will get the most recent version of that dataset in Kaggle. It uses the Kaggle api as well, which is a Godsend. Way better than what I was trying to do initially. 

I did my best to include informative comments so I wouldn't need to make this part too long but in short: 
This program establishes a connection to Kaggle, looks for a specific database, automatically makes a folder in the directory you have this .py file in, downloads the dataset into that folder, unzips it and all, and then reads it into the IDE so it can be manipulated any way we wish.